added test coverage for failed http requests
