+++ src/components/Sidebar/SideBar.test.tsx
@@ -3,9 +3,7 @@ import mockI18next from '@/mocks/mockI18next';
 import SideBar from './SideBar';
 import { renderWithProviders } from '@/utils/testUtils';
 import { INavigationItem } from '@/interfaces/common';
-import mockProjectCards from '@/mocks/mockProjectCards';
 import { setupStore } from '@/store';
-import { getProjectCardsThunk } from '@/reducers/projectCardSlice';
 import { matchExact } from '@/utils/common';
 import mockUsers from '@/mocks/mockUsers';
 import { getUsersThunk } from '@/reducers/authSlice';
@@ -28,9 +26,6 @@ describe('SideBar', () => {
   beforeEach(async () => {
     mockedAxios.get.mockResolvedValue(mockUsers);
     await store.dispatch(getUsersThunk());
-
-    mockedAxios.get.mockResolvedValue(mockProjectCards);
-    await store.dispatch(getProjectCardsThunk());
   });
 
   afterEach(async () => {
@@ -49,11 +44,6 @@ describe('SideBar', () => {
     navItems.forEach((n) => {
       expect(getByRole('button', { name: n.label })).toBeInTheDocument();
     });
-  });
-
-  it('populates projectCards to redux', async () => {
-    renderWithProviders(<SideBar />, { store });
-    expect(store.getState().projectCard.projectCards?.length).toBeGreaterThan(0);
   });
 
   // FIXME: tests with routes don't work yet
