+++ src/mocks/mockError.ts
@@ -1,0 +1,3 @@
+import { IError } from '@/interfaces/common';
+
+export const mockError: IError = { status: 400, message: 'Resource not found' };
